+++ models/motionTypeTemplates/PDFApplication.php
@@ -45,9 +45,7 @@ trait PDFApplication
         $initiatorSettings->contactEmail = InitiatorForm::CONTACT_REQUIRED;
         $type->supportTypeSettings       = json_encode($initiatorSettings, JSON_PRETTY_PRINT);
 
-        $settings                = new MotionType(null);
-        $settings->layoutTwoCols = 1;
-        $type->setSettingsObj($settings);
+        $type->setSettingsObj(new MotionType(null));
 
         $type->save();