+++ History.md
@@ -1,5 +1,9 @@
 # Version history
 
+### Version 4.0.2 [not released yet]
+
+- Bugfix: The PDF export of applications failed if a tabular data section was added in the motion, but no data rows were set.
+
 ### Version 4.0.1 (2018-09-02)
 
 - Bugfix: Motion comment tables were not created correctly when MySQL tables names are case-sensitive.